+++ src/constant.ts
@@ -1,9 +1,15 @@
 export const GPT_BUILDER_AI_KEY = "GPT_BUILDER_AI_KEY";
+export const GPT_BUILDER_AI_MODEl = "GPT_BUILDER_AI_MODEL";
 
 export const COMMAND = {
   clear: "builderAIAssistant.clearAll",
-  gptKey: "builderAIAssistant.gptKey",
+  gptKey: "builderAIAssistant.setting",
   selectFile: "builderAIAssistant.selectFile",
   openFile: "builderAIAssistant.openFile",
   requirement: "builderAIAssistant.enterRequirement"
 };
+
+export enum Model {
+  gpt_3_5 = "gpt-3.5-turbo",
+  gpt_4_0 = "gpt-4-turbo"
+}