+++ src/extension.ts
@@ -6,8 +6,8 @@ import * as path from "path";
 import { ExplorerItem, GptExplorerProvider } from "./fileExplorer";
 import { ChatGPT } from "./gpt";
 import { error } from "console";
-import { COMMAND } from "./constant";
-import { getGPTKey, setGPTKey } from "./utils";
+import { COMMAND, Model } from "./constant";
+import { getGPTKey, getGPTModel, setGPTKey, setGPTModel } from "./utils";
 
 let requirement: string | undefined;
 const converter = new showdown.Converter();
@@ -24,16 +24,21 @@ export function activate(context: vscode.ExtensionContext) {
   });
 
   vscode.commands.registerCommand(COMMAND.gptKey, async () => {
-    const gptKey = await vscode.window.showInputBox({
-      placeHolder: "Enter the GPT API Key",
-      password: true,
-      value: getGPTKey(context)
+    const options: {
+      [key: string]: (context: vscode.ExtensionContext) => Promise<void>;
+    } = {
+      "GPT Model": showQuickPick,
+      "GPT Key": showGPTInputBox
+    };
+    const quickPick = vscode.window.createQuickPick();
+    quickPick.items = Object.keys(options).map((label) => ({ label }));
+    quickPick.onDidChangeSelection((selection) => {
+      if (selection[0]) {
+        options[selection[0].label](context).catch(console.error);
+      }
     });
-
-    if (gptKey) {
-      setGPTKey(context, gptKey);
-      vscode.window.showInformationMessage(`GPT API Key stored successfully!`);
-    }
+    quickPick.onDidHide(() => quickPick.dispose());
+    quickPick.show();
   });
 
   vscode.commands.registerCommand(
@@ -83,13 +88,16 @@ export function activate(context: vscode.ExtensionContext) {
         );
         return;
       }
-
-      const gptOutput = await getGPTResponse(
-        context,
-        Array.from(provider.selectedFiles),
-        requirement
-      );
-      showOutput(context, gptOutput);
+      try {
+        const gptOutput = await getGPTResponse(
+          context,
+          Array.from(provider.selectedFiles),
+          requirement
+        );
+        showOutput(context, gptOutput);
+      } catch (err: any) {
+        vscode.window.showErrorMessage(err?.message);
+      }
     }
   });
 
@@ -101,25 +109,64 @@ export function activate(context: vscode.ExtensionContext) {
   );
 }
 
+export async function showQuickPick(context: vscode.ExtensionContext) {
+  const options: vscode.QuickPickItem[] = [
+    { label: "GPT 3.5", detail: Model.gpt_3_5 },
+    { label: "GPT 4.0", detail: Model.gpt_4_0 }
+  ];
+
+  const quickPick = vscode.window.createQuickPick();
+  let currentModel = getGPTModel(context);
+  quickPick.placeholder = currentModel
+    ? currentModel === Model.gpt_3_5
+      ? "GPT 3.5"
+      : "GPT 4.0"
+    : "GPT 3.5";
+  quickPick.canSelectMany = false;
+  quickPick.items = options;
+  quickPick.onDidChangeSelection((selections) => {
+    const selection = selections[0];
+    if (selection) {
+      setGPTModel(context, selection.detail ?? "");
+      vscode.window.showInformationMessage(
+        `You already selected ${selection.label}!`
+      );
+      quickPick.dispose();
+    } else {
+      vscode.window.showInformationMessage(`You don't select GPT Model!`);
+    }
+  });
+  quickPick.onDidHide(() => quickPick.dispose());
+  quickPick.show();
+}
+
+export async function showGPTInputBox(context: vscode.ExtensionContext) {
+  const gptKey = await vscode.window.showInputBox({
+    placeHolder: "Enter the GPT API Key",
+    password: true,
+    value: getGPTKey(context)
+  });
+
+  if (gptKey) {
+    setGPTKey(context, gptKey);
+    vscode.window.showInformationMessage(`GPT API Key stored successfully!`);
+  }
+}
+
 async function getGPTResponse(
   context: vscode.ExtensionContext,
   filePaths: string[],
   requirement: string
 ): Promise<string> {
-  try {
-    const apiKey = getGPTKey(context);
-    const gptAPI = new ChatGPT(apiKey ?? "");
-    const implement = await getFileContentImplement(filePaths);
-    const data = {
-      implement: implement,
-      requirement: requirement
-    };
-
-    return await gptAPI.requestAPI(data);
-  } catch {
-    vscode.window.showErrorMessage(error.toString());
-    throw error;
-  }
+  const apiKey = getGPTKey(context) ?? "";
+  const modelName = getGPTModel(context) ?? Model.gpt_3_5;
+  const gptAPI = new ChatGPT(apiKey, modelName);
+  const implement = await getFileContentImplement(filePaths);
+  const data = {
+    implement: implement,
+    requirement: requirement
+  };
+  return await gptAPI.requestAPI(data);
 }
 
 function showOutput(context: vscode.ExtensionContext, output: string) {